emergency-generator/generator_simulator/requirements.txt
Faker==13.3.4
paho-mqtt==1.6.1
python-dateutil==2.8.2
six==1.16.0
